analytics/gtm-stream-dataflow/requirements.txt
google-cloud-storage>=2.6.0,<3.0.0